+++ src/Text/Regex/Do/Pcre/Ascii/MatchHint.hs
@@ -1,23 +1,22 @@
 {- | this module uses
     <https://cdepillabout.github.io/haskell-type-families-presentation/#/ TypeFamilies>
 
-    this module is similar to "Text.Regex.Do.Pcre.Match". The differences are:
+    this module is similar to "Text.Regex.Do.Pcre.Ascii.Match". The differences are:
 
-    "Text.Regex.Do.Pcre.Match" is more flexible:
+    "Text.Regex.Do.Pcre.Ascii.Match" is more flexible:
         accepts 'Pattern' Regex,
         accepts 'Pattern' and 'Body' of different types
 
-    "Text.Regex.Do.Pcre.Match" needs to infer result type
+    "Text.Regex.Do.Pcre.Ascii.Match" needs to infer result type
 
     in this module the result type is determined by the hint
     -}
 
-{-# LANGUAGE TypeFamilies #-}
-module Text.Regex.Do.Pcre.MatchHint where
+module Text.Regex.Do.Pcre.Ascii.MatchHint where
 
 import Text.Regex.Do.Type.Do hiding (Once,All)
 import Text.Regex.PCRE.Wrap()
-import qualified Text.Regex.Do.Pcre.Match as M
+import qualified Text.Regex.Do.Pcre.Ascii.Match as M
 import Text.Regex.Do.Type.MatchHint
 import Data.ByteString
 
@@ -26,19 +25,10 @@ import Data.ByteString
 
     'Hint' and inferrable 'Pattern' or 'Body' type determine the instance
 
-    handy when working with 'OverloadedStrings', in other cases when compiler needs a hint
+    handy when working with 'OverloadedStrings', in other cases when compiler needs a hint  -}
 
-    >>> Test ("в"::ByteString) =~ "тихо в лесу"
 
-    True
 
-    >>> Once ("^all"::String) =~ "all the time"
-
-    \["all"\]
-
-    >>> PosLen' ("и"::String) =~ "бывает и хуже"
-
-    \[(13,2)\]      -}
 class (Hint hint, M.Match a a (F hint a)) =>
     MatchHint hint a where
     type F hint a
@@ -54,8 +44,13 @@ class (Hint hint, M.Match a a (F hint a)) =>
 instance MatchHint Test String where
     type F Test String = Bool
 
+
 instance MatchHint PosLen' String where
     type F PosLen' String = [PosLen]
+
+{- ^ >>> PosLen' ("и"::String) =~ "бывает и хуже"
+
+    \[(13,2)\]      -}
 
 instance MatchHint PosLen_ String where
     type F PosLen_ String = [[PosLen]]
@@ -63,11 +58,19 @@ instance MatchHint PosLen_ String where
 instance MatchHint Once String where
     type F Once String = [String]
 
+{- ^ >>> Once ("^all"::String) =~ "all the time"
+
+    \["all"\]       -}
+
 instance MatchHint All String where
     type F All String = [[String]]
 
 instance MatchHint Test ByteString where
     type F Test ByteString = Bool
+
+{- ^ >>> Test ("в"::ByteString) =~ "тихо в лесу"
+
+    True    -}
 
 instance MatchHint PosLen' ByteString where
     type F PosLen' ByteString = [PosLen]