+++ src/Text/Regex/Do/Pcre/Matchf.hs
@@ -2,26 +2,38 @@
 module Text.Regex.Do.Pcre.Matchf where
 
 import Text.Regex.Do.Type.Do
-import Text.Regex.Do.Type.Reexport
-import qualified Text.Regex.Do.Pcre.Result as R
-import qualified Text.Regex.Base.RegexLike as R hiding (makeRegex)
-import Text.Regex.Do.Type.Regex_
+import Text.Regex.Do.Type.Reexport as R
+import Text.Regex.Do.Result as R
+import Text.Regex.Base.RegexLike as R
+import Text.Regex.Do.Type.MatchHint
+
+type R_ b = R.RegexLike R.Regex b
 
 
-class (Rx_ a b, Functor f) => Matchf f a b where
-   marray_::Pattern a -> Body b -> f MatchArray
-   poslen_::Pattern a -> Body b -> f [PosLen]
-   poslen_ r0 b0 = R.poslen $ marray_ r0 b0
+class Matchf hint b where
+    type H hint
+    type P hint
+    marray_::R_ b => hint (Pattern R.Regex) -> Body b -> H hint
+    poslen_::R_ b => hint (Pattern R.Regex) -> Body b -> P hint
 
-instance Rx_ a b => Matchf Maybe a b where
-   marray_ p0 (Body b0) = R.matchOnce (r_ p0) b0
+instance Matchf Once b where
+    type H Once = Maybe MatchArray
+    type P Once = Maybe [PosLen]
+    marray_ (Once (Pattern p0)) (Body b0) = R.matchOnce p0 b0
+    poslen_ r0 b0 = R.poslen $ marray_ r0 b0
 
-instance Rx_ a b => Matchf [] a b where
-   marray_ p0 (Body b0) = R.matchAll (r_ p0) b0
+instance Matchf All b where
+    type H All = [MatchArray]
+    type P All = [[PosLen]]
+    marray_ (All (Pattern p0)) (Body b0) = R.matchAll p0 b0
+    poslen_ r0 b0 = R.poslen $ marray_ r0 b0
 
 
-once::Rx_ a b => Pattern a -> Body b -> [b]      -- ^ matched content
-once p0 b0 = maybe [] id $ R.allMatches b0 $ marray_ p0 b0
+once::(R_ b, R.Extract b) =>
+    Pattern R.Regex -> Body b -> [b]      -- ^ matched content
+once p0 b0 = maybe [] id $ R.allMatches b0 $ marray_ (Once p0) b0
 
-all::Rx_ a b => Pattern a -> Body b -> [[b]]       -- ^ matched content
-all p0 b0 = R.allMatches b0 $ marray_ p0 b0
+
+all::(R_ b, R.Extract b) =>
+    Pattern R.Regex -> Body b -> [[b]]       -- ^ matched content
+all p0 b0 = R.allMatches b0 $ marray_ (All p0) b0